+++ src/main/java/net/finmath/marketdata/model/curves/Curve.java
@@ -115,7 +115,7 @@ public class Curve extends AbstractCurve implements Serializable, Cloneable {
 		 * @param value The value (or y-value) of the point.
 		 * @param isParameter A boolean specifying if this point is considered a "degree of freedom", e.g., in a calibration.
 		 */
-		public Point(double time, double value, boolean isParameter) {
+		Point(double time, double value, boolean isParameter) {
 			super();
 			this.time = time;
 			this.value = value;
@@ -170,7 +170,7 @@ public class Curve extends AbstractCurve implements Serializable, Cloneable {
 		 * @throws CloneNotSupportedException Thrown, when the curve could not be cloned.
 		 */
 		public CurveBuilder(Curve curve) throws CloneNotSupportedException {
-			this.curve = (Curve)curve.clone();
+			this.curve = curve.clone();
 		}
 
 		/* (non-Javadoc)
@@ -475,7 +475,7 @@ public class Curve extends AbstractCurve implements Serializable, Cloneable {
 	@Override
 	public CurveInterface getCloneForParameter(double[] parameter) throws CloneNotSupportedException {
 		if(Arrays.equals(parameter, getParameter())) return this;
-		Curve newCurve = (Curve) this.clone();
+		Curve newCurve = this.clone();
 		newCurve.setParameterPrivate(parameter);
 
 		return newCurve;