+++ src/main/java/net/finmath/marketdata/model/curves/CurveFactory.java
@@ -28,6 +28,9 @@ public class CurveFactory {
 
 	private static DayCountConvention_ACT_365 modelDcc = new DayCountConvention_ACT_365();
 
+	private CurveFactory() {
+	}
+
 	/**
 	 * Creates a monthly index curve with seasonality and past fixings.
 	 * 
@@ -149,7 +152,7 @@ public class CurveFactory {
 			currentProjectedIndexValue /= seasonCurve.getValue(baseTime);
 
 			CurveInterface indexCurve = new IndexCurveFromDiscountCurve(name, currentProjectedIndexValue, discountCurve);
-			CurveInterface indexCurveWithSeason = new CurveFromProductOfCurves(name, referenceDate, new CurveInterface[] { indexCurve, seasonCurve });
+			CurveInterface indexCurveWithSeason = new CurveFromProductOfCurves(name, referenceDate, indexCurve, seasonCurve);
 			PiecewiseCurve indexCurveWithFixing = new PiecewiseCurve(indexCurveWithSeason, curveOfFixings, -Double.MAX_VALUE, fixingTimes[fixingTimes.length-1] + 1.0/365.0);
 			return indexCurveWithFixing;
 		}