+++ src/main/java/net/finmath/marketdata/model/curves/ForwardCurveNelsonSiegelSvensson.java
@@ -82,12 +82,11 @@ public class ForwardCurveNelsonSiegelSvensson extends AbstractCurve implements S
 
 	@Override
 	public double getForward(AnalyticModelInterface model, double fixingTime, double paymentOffset) {
-		paymentOffset = getPaymentOffset(fixingTime+periodOffset);
 		double daycountFraction = (paymentOffset*discountCurve.getTimeScaling());
 		if(daycountConvention != null) {
 			LocalDate fixingDate		= getDateFromModelTime(fixingTime+periodOffset);
-			LocalDate paymentDate		= getDateFromModelTime(fixingTime+periodOffset + paymentOffset);;
-			daycountFraction = Math.max(daycountConvention.getDaycountFraction(fixingDate, paymentDate), 1.0/365.0);
+			LocalDate paymentDate		= getDateFromModelTime(fixingTime+periodOffset + paymentOffset);
+            daycountFraction = Math.max(daycountConvention.getDaycountFraction(fixingDate, paymentDate), 1.0/365.0);
 		}
 
 		return (discountCurve.getDiscountFactor(model, fixingTime+periodOffset) / discountCurve.getDiscountFactor(model, fixingTime+periodOffset + paymentOffset) - 1.0) / daycountFraction;
@@ -99,16 +98,16 @@ public class ForwardCurveNelsonSiegelSvensson extends AbstractCurve implements S
 	}
 
 	@Override
-	public CurveBuilderInterface getCloneBuilder() throws CloneNotSupportedException {
+	public CurveBuilderInterface getCloneBuilder() {
 		return new CurveBuilderInterface() {
 			@Override
-			public CurveInterface build() throws CloneNotSupportedException {
+			public CurveInterface build() {
 				return ForwardCurveNelsonSiegelSvensson.this;
 			}
 
 			@Override
 			public CurveBuilderInterface addPoint(double time, double value, boolean isParameter) {
-				return this;
+				throw new UnsupportedOperationException("NSS curve does not support adding points.");
 			}			
 		};
 	}