+++ src/main/java/net/finmath/modelling/UnsupportedProduct.java
@@ -7,6 +7,7 @@
 package net.finmath.modelling;
 
 import java.io.Serializable;
+import java.util.Map;
 
 import net.finmath.marketdata.model.AnalyticModelInterface;
 import net.finmath.marketdata.products.AnalyticProductInterface;
@@ -18,11 +19,11 @@ import net.finmath.marketdata.products.AnalyticProductInterface;
  * 
  * @author Christian Fries
  */
-public class UnsupportedProduct implements Product, AnalyticProductInterface, Serializable {
+public class UnsupportedProduct implements ProductInterface, AnalyticProductInterface, Serializable {
 
 	private static final long serialVersionUID = 5375406324063846793L;
 	private final Exception exception;
-	
+
 	/**
 	 * Creates an unsupported product throwing an exception if its <code>getValue</code> method is called.
 	 * 
@@ -34,12 +35,17 @@ public class UnsupportedProduct implements Product, AnalyticProductInterface, Se
 	}
 
 	@Override
-	public Object getValue(double evaluationTime, Model model) {
+	public Object getValue(double evaluationTime, ModelInterface model) {
 		throw exception instanceof RuntimeException ? (RuntimeException)exception : new RuntimeException(exception);
 	}
 
 	@Override
 	public double getValue(double evaluationTime, AnalyticModelInterface model) {
+		throw exception instanceof RuntimeException ? (RuntimeException)exception : new RuntimeException(exception);
+	}
+
+	@Override
+	public Map<String, Object> getValues(double evaluationTime, ModelInterface model) {
 		throw exception instanceof RuntimeException ? (RuntimeException)exception : new RuntimeException(exception);
 	}