+++ src/main/java/net/finmath/montecarlo/AbstractMonteCarloProduct.java
@@ -9,10 +9,8 @@ import java.util.HashMap;
 import java.util.Map;
 
 import net.finmath.exception.CalculationException;
-import net.finmath.marketdata.model.AnalyticModelInterface;
-import net.finmath.modelling.Model;
-import net.finmath.modelling.Product;
-import net.finmath.modelling.ProductDescriptor;
+import net.finmath.modelling.ModelInterface;
+import net.finmath.modelling.ProductInterface;
 import net.finmath.stochastic.RandomVariableInterface;
 
 /**
@@ -20,7 +18,7 @@ import net.finmath.stochastic.RandomVariableInterface;
  * 
  * @author Christian Fries
  */
-public abstract class AbstractMonteCarloProduct<T extends ProductDescriptor> implements Product<T> {
+public abstract class AbstractMonteCarloProduct implements ProductInterface {
 
 	private final String currency;
 
@@ -34,10 +32,41 @@ public abstract class AbstractMonteCarloProduct<T extends ProductDescriptor> imp
 	}
 
 	@Override
-	public Object getValue(double evaluationTime, Model model) {
-		throw new IllegalArgumentException("The product " + this.getClass()
-				+ " cannot be valued against a model " + model.getClass() + "."
-				+ "It requires a model of type " + AnalyticModelInterface.class + ".");
+	public Object getValue(double evaluationTime, ModelInterface model) {
+		if(model instanceof MonteCarloSimulationInterface) {
+			try {
+				return getValue(evaluationTime, (MonteCarloSimulationInterface)model);
+			} catch (CalculationException e) {
+				return null;
+			}
+		}
+		else {
+			throw new IllegalArgumentException("The product " + this.getClass()
+			+ " cannot be valued against a model " + model.getClass() + "."
+			+ "It requires a model of type " + MonteCarloSimulationInterface.class + ".");
+		}
+	}
+
+	@Override
+	public Map<String, Object> getValues(double evaluationTime, ModelInterface model) {
+		Map<String, Object> results;
+		if(model instanceof MonteCarloSimulationInterface) {
+			try {
+				results = getValues(evaluationTime, (MonteCarloSimulationInterface)model);
+			} catch (CalculationException e) {
+				results = new HashMap<String, Object>();
+				results.put("exception", e);
+			}
+		}
+		else {
+			Exception e = new IllegalArgumentException("The product " + this.getClass()
+			+ " cannot be valued against a model " + model.getClass() + "."
+			+ "It requires a model of type " + MonteCarloSimulationInterface.class + ".");
+			results = new HashMap<String, Object>();
+			results.put("exception", e);
+		}
+		
+		return results;
 	}
 
 	/**