+++ src/test/java/net/finmath/marketdata/model/volatilities/CapletVolatilitiesParametricCalibrationTest.java
@@ -58,17 +58,17 @@ public class CapletVolatilitiesParametricCalibrationTest {
 	 * 
 	 * @return Array of parameters.
 	 */
-	@Parameters(name="{1}")
+	@Parameters(name="{0}")
 	public static Collection<Object[]> generateData()
 	{
 		return Arrays.asList(new Object[][] {
 				{ QuotingConvention.VOLATILITYLOGNORMAL },
 				{ QuotingConvention.VOLATILITYNORMAL },
 		});
-	};
+	}
 
-	
-	public CapletVolatilitiesParametricCalibrationTest(QuotingConvention calibrationTargetValueQuotingConvention) {
+
+    public CapletVolatilitiesParametricCalibrationTest(QuotingConvention calibrationTargetValueQuotingConvention) {
 		super();
 		this.calibrationTargetValueQuotingConvention = calibrationTargetValueQuotingConvention;
 		
@@ -238,13 +238,13 @@ public class CapletVolatilitiesParametricCalibrationTest {
 		 */
 		if(calibrationTargetValueQuotingConvention ==  QuotingConvention.VOLATILITYLOGNORMAL) {
 			// Check parameters
-			Assert.assertArrayEquals(new double[] { -1.04, 8.35 , 2.10, -0.01 }, calibratedParameters, 0.01 /* tolerance */ );
+			Assert.assertArrayEquals(new double[] { -1.04, 8.37 , 2.11, -0.01 }, calibratedParameters, 0.01 /* tolerance */ );
 			// Check rms (not a very strict test)
 			Assert.assertTrue(rms < 0.08);
 		}
 		else {
 			// Check parameters
-			Assert.assertArrayEquals(new double[] { -1.08, 8.73 , 2.19, -0.01 }, calibratedParameters, 0.01 /* tolerance */ );
+			Assert.assertArrayEquals(new double[] { -1.08, 8.76 , 2.19, -0.01 }, calibratedParameters, 0.01 /* tolerance */ );
 			// Check rms (not a very strict test)
 			Assert.assertTrue(rms < 0.02);
 		}